+++ tests/Unit/DataStructures/CMakeLists.txt
@@ -7,6 +7,7 @@ set(LIBRARY_SOURCES
   Test_DataBox.cpp
   Test_DataBoxTag.cpp
   Test_DataVector.cpp
+  Test_Deferred.cpp
   Test_GeneralIndexIterator.cpp
   Test_Identity.cpp
   Test_Index.cpp
